requirements.txt
Flask==3.1.0
bg-remover==0.1
pillow==11.0.0
werkzeug==3.1.3
